Migrate interview to sapphire (#175)
* merge popped commit

* migrate interviewer commands
