+++ package.json
@@ -22,6 +22,8 @@
     "@sapphire/decorators": "^4.2.3",
     "@sapphire/framework": "^2.4.0",
     "@sapphire/plugin-logger": "^2.1.3",
+    "@sapphire/plugin-subcommands": "^2.1.3",
+    "@sapphire/utilities": "^3.3.0",
     "concurrently": "^6.1.0",
     "cron": "^1.8.2",
     "discord.js": "^13.6.0",
