+++ src/commands/coffeeChat/coffeeChatTest.ts
@@ -1,14 +1,14 @@
 import { Message, MessageEmbed } from 'discord.js';
 import { Args, Command, CommandOptions } from '@sapphire/framework';
 import { ApplyOptions } from '@sapphire/decorators';
-import { container } from '@sapphire/pieces';
 import { testPerformance } from '../../components/coffeeChat';
 import { EMBED_COLOUR } from '../../utils/embeds';
+import { BOT_PREFIX } from '../../bot';
 
 @ApplyOptions<CommandOptions>({
   aliases: ['coffee-test', 'coffeetest', 'coffee-chat-test'],
   description: 'Tests coffeechatmatch, displaying the number of rounds before there is a duplicate.',
-  detailedDescription: `**Examples:**\n\`${container.client.options.defaultPrefix}coffeetest 10\``,
+  detailedDescription: `**Examples:**\n\`${BOT_PREFIX}coffeetest 10\``,
   requiredUserPermissions: ['ADMINISTRATOR']
 })
 export class CoffeeChatTestCommand extends Command {
